Move post title to own component
